requirements.txt
Flask==2.3.3
authlib==1.2.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
requests==2.31.0
Werkzeug==2.3.7
python-dotenv==1.0.0
boto3==1.34.69
gunicorn==21.2.0
